action: Update to actions/upload-artifact@v4
Include report_title_suffix because upload-artifact@v4 doesn't allow
merging of multiple inputs from distinct steps into the same artifact.

Use gh to download artifacts because actions/download-artifact@v4 is
hostile to many use cases. (It doesn't provide an output if it can't
find an artifact, and it doesn't have a way not to fail if it can't
find an artifact.)

Enable apply.pl to retrieve a single v4 matrix artifact.