Support using a cached jq
GitHub hosted runners include jq so this isn't necessary.
In cases where jq isn't present, it will be installed in a fixed
location, if that location persists, it'll be reused.

Note that while the GitHub runtime does support caching, using that
feature in public repositories is not recommended at this time.

Supporting cached jq is mostly designed for standalone consumers of
check-spelling.