unknown-words: Avoid modifying original metadata files
get_project_files makes new files available to consumers and uses cleanup_file
to clean files.

Part of cleanup_file wants to complain about original content,
so it needs the original paths.

The rest is involved in constructing the content as we want to consume it.

As such, it is ok to construct the content to consume in the new file
instead of the original file.