action: Use local actions
Instead of eagerly downloading all action dependencies, we reference
them using local action stubs. The local action stubs reference the real
actions.

The real actions are only downloaded when the stub actions are used by
steps.

This means that if we, e.g., don't use upload-sarif, we don't have to
wait 5-6 seconds to download 300mb of data.

See:
https://api.github.com/repos/github/codeql-action/issues/2542

This change also means that only the real or fallback action will be
downloaded instead of downloading both.

---

For some reason, using inputs.GITHUB_TOKEN with a default of
`github.token`, and then using a stub action with boolean logic resulted
in the token getting lost. Instead, use github.token as an explicit
fallback.